automatically backed up by learn

+++ currency.rb
@@ -31,7 +31,7 @@ def aud_to_usd(aud)
 end
   
 puts "How much money do you want to convert?"
-amount = gets.chomp().to_i
+amount = gets.chomp().to_f
  
 puts "Pick from the following menu:"
 puts "1. USD to EUR"
@@ -43,7 +43,7 @@ puts "6. JPY to USD"
 puts "7. GBP to USD"
 puts "8. AUD to USD"
  
-choice = gets.chomp().to_i
+choice = gets.chomp().to_f
  
 case choice
 when 1
